requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
scapy==2.4.4
gunicorn==20.0.4
google-cloud-storage==1.31.0
